DETR_camera/requirements.txt
cython
git+https://github.com/cocodataset/cocoapi.git#subdirectory=PythonAPI&egg=pycocotools
submitit
torch>=1.5.0
torchvision>=0.6.0
git+https://github.com/cocodataset/panopticapi.git#egg=panopticapi
scipy
onnx
onnxruntime
requests
matplotlib
Pillow
ipywidgets
IPython
opencv-contrib-python
numpy
